requirements_docs.txt
pypandoc==1.8.1
matplotlib==3.5.2
vtk<9.1.0
pyvista==0.35.1
imageio==2.19.3
imageio-ffmpeg==0.4.7
Sphinx==5.0.2
nbsphinx==0.8.9
sphinx-autobuild==2021.3.14
sphinxcontrib-napoleon==0.7
sphinxcontrib-websupport==1.2.4
sphinx_design
pytest-sphinx==0.4.0
sphinx-notfound-page==0.8
sphinx-copybutton==0.5.0
sphinx-gallery==0.10.1
ansys_sphinx_theme==0.4.2
graphviz
enum-tools[sphinx]
